+++ src/project/build.task.ts
@@ -4,7 +4,6 @@ import {spawnProcess} from '../utils/process.js';
 import {Timings} from '../utils/time.js';
 import {buildSourceDirectory} from '../build/buildSourceDirectory.js';
 import {loadGroConfig} from '../config/config.js';
-import {configureLogLevel} from '../utils/log.js';
 import {clean} from '../fs/clean.js';
 
 export const task: Task = {
@@ -15,7 +14,6 @@ export const task: Task = {
 
 		const timeToLoadConfig = timings.start('load config');
 		const config = await loadGroConfig(dev);
-		configureLogLevel(config.logLevel);
 		timeToLoadConfig();
 
 		// build everything using the normal build process - js files will compiled again by `tsc` later